requirements.txt
spire.doc>=11.0.0
spire.xls>=11.0.0
pywin32>=306
pandas>=1.5.0
openpyxl>=3.0.0
pywin32>=306
pywebview>=4.0.0
Pillow>=9.0.0
PyYAML>=6.0
tkcalendar>=1.6.1
PyPDF2>=3.0.0
reportlab>=3.6.0 